Refactored code using unityevent
