+++ ferrous-game/Assets/Scripts/Mechanics/OpenDoor.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 7a102b61cad1a3341a1e1156be5ffd02
+guid: 763af9684fe923a41ae3e5f58b0ed83c
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
